front-back order

+++ RealEstateBackend/Controllers/AddressesController.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using RealEstate.Models.Domains;
@@ -31,40 +32,82 @@ namespace RealEstate.Controllers
         }
 
 
+        [HttpGet]
+        [Route("Buyer")]
+        [Authorize(Roles = "Buyer")]
+        public async Task<IActionResult> GetAllByBuyer()
+        {
+            string buyerIdStr = User.FindFirst("userId")?.Value;
+
+            if (!int.TryParse(buyerIdStr, out int buyerId))
+            {
+                return Unauthorized("Buyer not found.");
+            }
+
+
+            var addresses = await _addressRepository.GetAllByBuyerAsync(buyerId);
+            var result = _mapper.Map<List<AddressDto>>(addresses);
+            return Ok(result);
+        }
+
+
         // GET: api/Address/5
         [HttpGet("{id}")]
+        [Authorize(Roles = "Buyer")]
         public async Task<IActionResult> GetById(int id)
         {
+            string buyerIdStr = User.FindFirst("userId")?.Value;
+
+            if (!int.TryParse(buyerIdStr, out int buyerId))
+            {
+                return Unauthorized("Buyer not found.");
+            }
+
             var address = await _addressRepository.GetByIdAsync(id);
             if (address == null)
                 return NotFound("Address not found!");
 
-            var dto = _mapper.Map<CreateAddressDto>(address); 
+            if (address.BuyerId != buyerId)
+            {
+                return Unauthorized("Not allowed to delete this address.");
+            }
+
+            var dto = _mapper.Map<AddressDto>(address); 
             return Ok(dto);
         }
 
 
         [HttpPost]
+        [Authorize(Roles = "Buyer")]
         public async Task<IActionResult> Create([FromBody] CreateAddressDto dto)
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
-            // Check if the buyer exists
-            var buyer = await _buyerRepository.GetByIdAsync(dto.BuyerId);
+
+            string buyerIdStr = User.FindFirst("userId")?.Value;
+
+            if (!int.TryParse(buyerIdStr, out int buyerId))
+            {
+                return Unauthorized("Buyer not found.");
+            }
+
+            var buyer = await _buyerRepository.GetByIdAsync(buyerId);
             if (buyer == null)
             {
-                return NotFound($"Buyer with ID {dto.BuyerId} not found.");
+                return NotFound($"Buyer with ID {buyerId} not found.");
             }
 
             var address = _mapper.Map<Address>(dto); // Address is your domain model
-            await _addressRepository.CreateAsync(address);
+            address.BuyerId = buyerId;
+            address = await _addressRepository.CreateAsync(address);
             var resultDto = _mapper.Map<AddressDto>(address);
 
-            return CreatedAtAction(nameof(GetById), new { id = address.Id }, resultDto);
+            return Ok(resultDto);
         }
 
         // PUT: api/Address/5
         [HttpPut("{id}")]
+        [Authorize(Roles = "Buyer")]
         public async Task<IActionResult> Update(int id, [FromBody] UpdateAddressDto updateDto)
         {
             if (!ModelState.IsValid)
@@ -83,11 +126,25 @@ namespace RealEstate.Controllers
 
         // DELETE: api/Address/5
         [HttpDelete("{id}")]
+        [Authorize(Roles = "Buyer")]
         public async Task<IActionResult> Delete(int id)
         {
+            string buyerIdStr = User.FindFirst("userId")?.Value;
+
+            if (!int.TryParse(buyerIdStr, out int buyerId))
+            {
+                return Unauthorized("Buyer not found.");
+            }
+
             var address = await _addressRepository.GetByIdAsync(id);
             if (address == null)
                 return NotFound("Address not found!");
+
+            if(address.BuyerId != buyerId)
+            {
+                return Unauthorized("Not allowed to delete this address.");
+
+            }
 
             await _addressRepository.DeleteAsync(id);
             return Ok(new { message = "Address deleted successfully." });
